+++ src/query/expression/src/row_encoding.rs
@@ -12,11 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use databend_common_column::bitmap::Bitmap;
 use databend_common_column::types::months_days_micros;
 
-use super::row_converter::null_sentinel;
-use crate::types::binary::BinaryColumnBuilder;
 use crate::types::i256;
 use crate::types::F32;
 use crate::types::F64;
@@ -104,61 +101,5 @@ impl FixedLengthEncoding for months_days_micros {
 
     fn encode(self) -> [u8; 8] {
         self.total_micros().encode()
-    }
-}
-
-pub fn encode<T, I>(
-    out: &mut BinaryColumnBuilder,
-    iter: I,
-    (all_null, validity): (bool, Option<&Bitmap>),
-    asc: bool,
-    nulls_first: bool,
-) where
-    T: FixedLengthEncoding,
-    I: IntoIterator<Item = T>,
-{
-    if all_null {
-        for offset in out.offsets.iter_mut().skip(1) {
-            let start = *offset as usize;
-            let end = start + T::ENCODED_LEN;
-            out.data[start] = null_sentinel(nulls_first);
-            *offset = end as u64;
-        }
-        return;
-    }
-
-    let Some(validity) = validity else {
-        for (offset, val) in out.offsets.iter_mut().skip(1).zip(iter) {
-            let start = *offset as usize;
-            let end = start + T::ENCODED_LEN;
-            let to_write = &mut out.data[start..end];
-            to_write[0] = 1;
-            let mut encoded = val.encode();
-            if !asc {
-                // Flip bits to reverse order
-                encoded.as_mut().iter_mut().for_each(|v| *v = !*v)
-            }
-            to_write[1..].copy_from_slice(encoded.as_ref());
-            *offset = end as u64;
-        }
-        return;
-    };
-
-    for ((offset, val), v) in out.offsets.iter_mut().skip(1).zip(iter).zip(validity) {
-        let start = *offset as usize;
-        let end = start + T::ENCODED_LEN;
-        if v {
-            let to_write = &mut out.data[start..end];
-            to_write[0] = 1;
-            let mut encoded = val.encode();
-            if !asc {
-                // Flip bits to reverse order
-                encoded.as_mut().iter_mut().for_each(|v| *v = !*v)
-            }
-            to_write[1..].copy_from_slice(encoded.as_ref());
-        } else {
-            out.data[start] = null_sentinel(nulls_first);
-        }
-        *offset = end as u64;
     }
 }
