+++ src/query/pipeline/transforms/src/processors/transforms/sort/rows.rs
@@ -12,49 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod common;
-mod simple;
-mod utils;
-
 use std::fmt::Debug;
 use std::ops::Range;
 
-pub use common::*;
 use databend_common_exception::ErrorCode;
 use databend_common_exception::Result;
 use databend_common_expression::types::ArgType;
 use databend_common_expression::types::DataType;
-use databend_common_expression::BlockEntry;
 use databend_common_expression::Column;
-use databend_common_expression::DataBlock;
-use databend_common_expression::DataSchemaRef;
 use databend_common_expression::Scalar;
-use databend_common_expression::SortColumnDescription;
-pub use simple::*;
-pub use utils::*;
-
-/// Convert columns to rows.
-pub trait RowConverter<T: Rows>
-where Self: Sized + Debug
-{
-    fn create(
-        sort_columns_descriptions: &[SortColumnDescription],
-        output_schema: DataSchemaRef,
-    ) -> Result<Self>;
-    fn convert(&self, columns: &[BlockEntry], num_rows: usize) -> Result<T>;
-
-    fn convert_data_block(
-        &self,
-        sort_desc: &[SortColumnDescription],
-        data_block: &DataBlock,
-    ) -> Result<T> {
-        let order_by_cols = sort_desc
-            .iter()
-            .map(|desc| data_block.get_by_offset(desc.offset).clone())
-            .collect::<Vec<_>>();
-        self.convert(&order_by_cols, data_block.num_rows())
-    }
-}
 
 /// Rows can be compared.
 pub trait Rows
