+++ src/query/pipeline/transforms/tests/it/main.rs
@@ -13,3 +13,4 @@
 // limitations under the License.
 
 mod merger;
+mod row_converter;
